requirements.txt
asgiref==3.2.4
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-redis==4.11.0
django-redis-cache==2.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
